app/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
djangorestframework==3.15.2
markdown==3.6
django-filter==24.3
django-cors-headers==4.4.0
django-phone-field==1.8.1